Move factfilter.evaluator tests to core.interpreter package
